+++ src/pages/AsesoriaJuridicaDetalle/Seguimiento/Actuaciones/Actuacion/Demanda.js
@@ -5,6 +5,7 @@ import Spin from "../../../../../components/Spin";
 import { toast } from "react-toastify";
 import API from "utils/Axios";
 import Footer from "./Footer";
+import { CONFIRM_BORRAR_ARCHIVO } from "constants/apiContants";
 
 const Demanda = ({
   actuacion,
@@ -14,16 +15,17 @@ const Demanda = ({
   persona = "",
 }) => {
   const borrarArchivo = async (archivo) => {
-    console.log(archivo);
-    setCargando(true);
-    try {
-      await API.delete(`asesorias/docsanexos/${archivo.id}`);
-      toast.success("Archivo borrado");
-      setCargando(false);
-      anexoBorrado(archivo.id);
-    } catch (error) {
-      setCargando(false);
-      toast.error(error.toString());
+    if (window.confirm(CONFIRM_BORRAR_ARCHIVO)) {
+      setCargando(true);
+      try {
+        await API.delete(`asesorias/docsanexos/${archivo.id}`);
+        toast.success("Archivo borrado");
+        setCargando(false);
+        anexoBorrado(archivo.id);
+      } catch (error) {
+        setCargando(false);
+        toast.error(error.toString());
+      }
     }
   };
   const [cargando, setCargando] = useState(false);
